requirements.txt
opencv-python == 4.6.0.66
opencv-contrib-python == 4.6.0.66
pandas == 1.5.0
scikit-learn == 1.1.2
matplotlib == 3.6.0
scikit-image == 0.19.3
seaborn == 0.12.0
tqdm == 4.64.1
SimpleITK-SimpleElastix
PyYAML == 6.0
unzip == 1.0.0
antspyx == 0.4.2
monai